feat(seo): per-locale <html lang>, canonical, x-default, noindex on fallbacks
Restructure app/ into (en) and (zh) route groups, each with its own root
layout that owns <html lang>. Server-rendered HTML now correctly carries
lang="en" for /* and lang="zh" for /zh/*. Resolves the audit's P0 item
where Chinese pages were indexed as English because all routes inherited
<html lang="en"> from the single root layout.

Changes:
- app/(en)/layout.tsx and app/(zh)/layout.tsx as parallel root layouts.
  Both own <html>, <body>, theme script, ThemeProvider, LanguageProvider
  (with initialLanguage), header/footer.
- Old app/layout.tsx and app/zh/layout.tsx removed. Their metadata moved
  into the new root layouts.
- All EN routes moved into app/(en)/. All ZH routes moved into
  app/(zh)/zh/. sitemap.ts, robots.ts, globals.css stay at app/ root.
- LanguageProvider now receives initialLanguage from the layout so the
  SSR pass renders the correct language without waiting for the client
  effect to re-sync from URL.

Metadata fixes on every detail and index page:
- alternates.canonical set on every page with stable content.
- alternates.languages now includes x-default everywhere.
- Detail pages where essay/periodic/series lang doesn't match the URL
  locale (the cross-language fallback / meta-refresh case) return
  robots: { index: false, follow: true } so they don't dilute the index.
- ZH about page description: replace "杨飞同" with "Feitong Yang".

Tests:
- layout.test.ts now covers both root layouts and also asserts
  metadataBase, x-default, and template/default structure.
- home-page.test.tsx import path updated for the move.

164 tests pass. Build OK under output: 'export'. Dev server smoke-tested:
EN routes serve <html lang="en">, ZH routes serve <html lang="zh"> in
the initial server HTML.
