+++ apps/blog/__tests__/home-page.test.tsx
@@ -4,7 +4,7 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import HomePage from '@/app/page';
+import HomePage from '@/app/(en)/page';
 
 describe('HomePage', () => {
   it('renders the page title', () => {
