+++ apps/blog/app/(en)/about/page.tsx
@@ -1,0 +1,19 @@
+import type { Metadata } from 'next';
+import { AboutPage } from '@/components/pages';
+
+export const metadata: Metadata = {
+  title: 'About',
+  description: 'About Feitong Yang - career, background, and interests.',
+  alternates: {
+    canonical: '/about',
+    languages: {
+      en: '/about',
+      zh: '/zh/about',
+      'x-default': '/about',
+    },
+  },
+};
+
+export default function Page() {
+  return <AboutPage language="en" />;
+}
